refactor: #459 follow 관련 아키텍처 통일
