--- Spoony-iOS/Spoony-iOS/Network/DependencyKey/DetailUseCaseKey.swift
@@ -1,13 +1,0 @@
-//
-//  DetailUseCaseKey.swift
-//  Spoony-iOS
-//
-//  Created by 최안용 on 5/4/25.
-//
-
-import Dependencies
-
-enum PostUseCaseKey: DependencyKey {
-    static let liveValue: PostUseCase = PostUseCaseImpl()
-    static let testValue: PostUseCase = MockPostUseCase()
-}
